+++ lib/widgets/search/task_status_filter.dart
@@ -5,8 +5,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:lotti/blocs/journal/journal_page_cubit.dart';
 import 'package:lotti/blocs/journal/journal_page_state.dart';
-
-import 'filter_choice_chip.dart';
+import 'package:lotti/widgets/search/filter_choice_chip.dart';
+import 'package:quiver/collection.dart';
 
 class TaskStatusFilter extends StatefulWidget {
   const TaskStatusFilter({super.key});
@@ -28,6 +28,7 @@ class _TaskStatusFilterState extends State<TaskStatusFilter> {
             lineSpacing: 5,
             children: [
               ...snapshot.taskStatuses.map(TaskStatusChip.new),
+              const TaskStatusAllChip(),
             ],
           ),
         );
@@ -60,14 +61,55 @@ class TaskStatusChip extends StatelessWidget {
 
     return BlocBuilder<JournalPageCubit, JournalPageState>(
       builder: (context, snapshot) {
+        final cubit = context.read<JournalPageCubit>();
+
         void onTap() {
-          context.read<JournalPageCubit>().toggleSelectedTaskStatus(status);
+          cubit.toggleSelectedTaskStatus(status);
+          HapticFeedback.heavyImpact();
+        }
+
+        void onLongPress() {
+          cubit.selectSingleTaskStatus(status);
           HapticFeedback.heavyImpact();
         }
 
         return FilterChoiceChip(
           label: '${localizationLookup[status]}',
           isSelected: snapshot.selectedTaskStatuses.contains(status),
+          onTap: onTap,
+          onLongPress: onLongPress,
+        );
+      },
+    );
+  }
+}
+
+class TaskStatusAllChip extends StatelessWidget {
+  const TaskStatusAllChip({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocBuilder<JournalPageCubit, JournalPageState>(
+      builder: (context, snapshot) {
+        final cubit = context.read<JournalPageCubit>();
+
+        final isSelected = setsEqual(
+          snapshot.selectedTaskStatuses,
+          snapshot.taskStatuses.toSet(),
+        );
+
+        void onTap() {
+          if (isSelected) {
+            cubit.clearSelectedTaskStatuses();
+          } else {
+            cubit.selectAllTaskStatuses();
+          }
+          HapticFeedback.heavyImpact();
+        }
+
+        return FilterChoiceChip(
+          label: 'ALL',
+          isSelected: isSelected,
           onTap: onTap,
         );
       },